update flash players to use relative base urls, backend_wrap
with this update, all the SWFs that request video have been updated
to use relative URLs (or specify the base url in flashvars, depending
on the player).

this removes the need to patch all SWFs.

additionally:
- a backend_wrap.js file was added to /back/ made to automatically
bring yt2009 back up after a failure. it's made with instances that
are supposed to be up 24/7 in mind.
to take advantage of it, use the command:

node backend_wrap.js
instead of
node backend.js

to start yt2009.

- the "public" config option will no longer do anything. publishing
your instance to the instance list is no longer possible straight
from the yt2009 code.
the instance list is set to shut down in late-march.