adaptive_old, bugfixes, nonhd stream while download, speedups
an adaptive_old flag was added globally as an addition
to only_old. it will follow the current date and month
and internally adjust channel and search only_old
to those dates.

this will result in a flow of "new" videos as they
were uploaded in the picked year.

additionally:
- fixed user avatars not showing correctly on some watchpages,
mostly videos by music related channels.
- fixed #249, #250, #259, #260 (huge thanks, sh20000sh!)
- fixed a bug where videos with 0 views could show NaN
in some places.
- non-hd video will now be streamed while downloading
on supported setups.
- fixed statistics & data tab possibly showing future dates
as dates of events happening.
- video url on watchpage was corrected to show the leading "www."
(thanks, RF64!)
- lowered clean watchpage load times by 0.5s on average
- lowered clean channel load times by 1.5s on average
- a "disable_old_banners" channel flag was added that will
not wait for old banner and background on a channel page.
will further reduce loading times by 1s on average.

// [!!] this update will overwrite existing channel caches,
// as the old format is incompatible. backup your cache
// if used to store historical data.